refactor(videoroom): msg_options -> params
