+++ jarust_plugins/src/video_room/mod.rs
@@ -1,5 +1,5 @@
 pub mod events;
 pub mod handle;
 pub mod jahandle_ext;
-pub mod msg_options;
+pub mod params;
 pub mod responses;
